Transferring startup logic to run.sh file and added installer sh file.
Also added new stuff to nix. IDK if they are working lol. They seem to be but I am on arch using the nix os package manager to test this and idk if it is even doing its thing like its supposed to cause I see a window but why do I get egl errors